ref: make common test mod
