+++ node/tests/new_beacon.rs
@@ -14,9 +14,20 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkOS library. If not, see <https://www.gnu.org/licenses/>.
 
+mod common;
+use common::TestPeer;
+
+use snarkos_node::new_beacon::Beacon;
+use snarkos_node_executor::NodeType;
+use snarkvm::prelude::{ConsensusMemory, PrivateKey, Testnet3 as CurrentNetwork};
+
+use std::str::FromStr;
+
+use pea2pea::Pea2Pea;
+
 #[tokio::test]
 async fn handshake_responder_side() {
-    // 1. Create beacon instance.
+    // Create a beacon instance.
     let beacon = Beacon::<CurrentNetwork, ConsensusMemory<CurrentNetwork>>::new(
         "127.0.0.1:4133".parse().unwrap(),
         None,
@@ -28,127 +39,14 @@ async fn handshake_responder_side() {
     .await
     .expect("couldn't create beacon instance");
 
+    // Spin up a test peer.
     let peer = TestPeer::new(NodeType::Validator).await;
 
+    // Verify the handshake works when the peer initates a connection with the node.
     assert!(
         peer.node()
             .connect(beacon.router().network().listening_addr().expect("beacon listener should exist"))
             .await
             .is_ok()
     );
-
-    // 2. Spin up test peer.
-    // 3. Connect the peer to the beacon, assert on handshake.
-}
-
-use snarkos_node::new_beacon::{Beacon, PeerMeta};
-use snarkos_node_executor::{NodeType, RawStatus, Status};
-use snarkos_node_messages::{
-    ChallengeRequest,
-    ChallengeResponse,
-    Data,
-    Disconnect,
-    DisconnectReason,
-    Message,
-    MessageCodec,
-};
-use snarkvm::prelude::{Block, ConsensusMemory, FromBytes, Network, PrivateKey, Testnet3 as CurrentNetwork};
-
-use std::{
-    io,
-    net::{IpAddr, Ipv4Addr, SocketAddr},
-    str::FromStr,
-};
-
-use futures_util::{sink::SinkExt, TryStreamExt};
-use pea2pea::{protocols::Handshake, Config, Connection, ConnectionSide, Node, Pea2Pea};
-use tokio_util::codec::Framed;
-
-const ALEO_MAXIMUM_FORK_DEPTH: u32 = 4096;
-
-#[derive(Clone)]
-struct TestPeer {
-    node: Node,
-    node_type: NodeType,
-}
-
-impl Pea2Pea for TestPeer {
-    fn node(&self) -> &Node {
-        &self.node
-    }
-}
-
-impl TestPeer {
-    pub async fn new(node_type: NodeType) -> Self {
-        let peer = Self {
-            node: Node::new(Config {
-                listener_ip: Some(IpAddr::V4(Ipv4Addr::LOCALHOST)),
-                max_connections: 200,
-                ..Default::default()
-            })
-            .await
-            .expect("couldn't create test peer"),
-            node_type,
-        };
-
-        peer.enable_handshake().await;
-        //  client.enable_reading().await;
-        //  client.enable_writing().await;
-        //  client.enable_disconnect().await;
-
-        peer
-    }
-
-    pub fn node_type(&self) -> NodeType {
-        self.node_type
-    }
-}
-
-#[async_trait::async_trait]
-impl Handshake for TestPeer {
-    async fn perform_handshake(&self, mut conn: Connection) -> io::Result<Connection> {
-        let local_addr = self.node().listening_addr().expect("listening address should be present");
-
-        let stream = self.borrow_stream(&mut conn);
-        let mut framed = Framed::new(stream, MessageCodec::<CurrentNetwork>::default());
-
-        // TODO (howardwu): Make this step more efficient (by not deserializing every time).
-        // Retrieve the genesis block header.
-        let genesis_header = *Block::<CurrentNetwork>::from_bytes_le(CurrentNetwork::genesis_bytes())
-            .expect("genesis block bytes should be valid")
-            .header();
-
-        // Send a challenge request to the peer.
-        let message = Message::<CurrentNetwork>::ChallengeRequest(ChallengeRequest {
-            version: Message::<CurrentNetwork>::VERSION,
-            fork_depth: ALEO_MAXIMUM_FORK_DEPTH,
-            node_type: self.node_type(),
-            status: Status::Peering,
-            listener_port: local_addr.port(),
-        });
-        framed.send(message).await?;
-
-        // Receive the challenge request.
-        let _challenge_request = framed.try_next().await?.unwrap();
-
-        // TODO(nkls): add assertions on the contents.
-
-        // Send the challenge response.
-        let message = Message::ChallengeResponse(ChallengeResponse { header: Data::Object(genesis_header) });
-        framed.send(message).await?;
-
-        // Receive the challenge response.
-        let Message::ChallengeResponse(challenge_response) = framed.try_next().await.unwrap().unwrap() else {
-            panic!("didn't get challenge response")
-        };
-
-        // Perform the deferred non-blocking deserialization of the block header.
-        let Ok(block_header) = challenge_response.header.deserialize().await else {
-            panic!("block header not present")
-        };
-
-        assert_eq!(block_header, genesis_header);
-
-        Ok(conn)
-    }
 }
